feat: OnAdventureViewmodel 재구현
